requirements.txt
numpy
matplotlib
pyserial
cryptography
pyjwt
httpx
httpx[http2]
scipy